requirements.txt
streamlit==1.12.0
Pillow==7.0.0
requests==2.22.0
streamlit-lottie== 0.0.3

